remove extra module

+++ src/app/customer/quib-user/quib-user.component.ts
@@ -5,14 +5,14 @@ import { NgxUiLoaderService, SPINNER } from 'ngx-ui-loader';
 import { TABLE_HEADING } from '../../_models/table_heading'
 import { Table } from 'primeng/table';
 import { ToastrMsgService } from 'src/app/_services/toastr-msg.service';
-import { PrimeNGConfig } from 'primeng/api';
-import {ConfirmationService, ConfirmEventType, MessageService} from 'primeng/api';
+import { ConfirmationService} from 'primeng/api';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 
 @Component({
   selector: 'app-quib-user',
   templateUrl: './quib-user.component.html',
-  styleUrls: ['./quib-user.component.scss']
+  styleUrls: ['./quib-user.component.scss'],
+  providers: [ConfirmationService]
 })
 export class QuibUserComponent implements OnInit {
   @ViewChild('dt') dt: Table | undefined;
@@ -27,10 +27,8 @@ export class QuibUserComponent implements OnInit {
     private ngxLoader: NgxUiLoaderService,
     private toastr: ToastrMsgService,
     private fb: FormBuilder,
-    private primengConfig: PrimeNGConfig,
     private confirmationService: ConfirmationService,
-     private messageService: MessageService
-  ) { 
+  ) {
     this.quibUserForm = this.fb.group({
       displayName: ["", [Validators.required]],
       firstName: ['', [Validators.required]],
@@ -46,12 +44,11 @@ export class QuibUserComponent implements OnInit {
   }
 
   ngOnInit(): void {
-    this.primengConfig.ripple = true;
     this.sidebarSpacing = 'contracted';
     this.fgsType = SPINNER.squareLoader
     this.ngxLoader.start();
     this.sidebarSpacing = 'contracted';
-     this.getOrderList()
+    this.getOrderList()
     this.cols = [
       { field: 'displayName', show: true, headers: 'Display Name' },
       { field: 'firstName', show: true, headers: 'First Name' },
@@ -63,7 +60,7 @@ export class QuibUserComponent implements OnInit {
       { field: 'FRS', show: true, headers: 'FRS' },
       { field: 'UNP', show: true, headers: 'UNP' },
       { field: 'status', show: true, headers: 'Status' },
-      
+
     ]
   }
 
@@ -78,23 +75,31 @@ export class QuibUserComponent implements OnInit {
     this.dt.filterGlobal(($event.target as HTMLInputElement).value, stringVal);
   }
   deleteUser(order: number) {
-    this.ngxLoader.start();
-    this.QuibService.deleteUser(order).subscribe(res => {
-      if (res) {
-        this.toastr.showSuccess("user deleted successfully", "user delete")
-        this.getOrderList()
-      }
-    })
+    this.confirmationService.confirm({
+      message: 'Are you sure that you want to delete user?',
+      header: 'Confirmation',
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.ngxLoader.start();
+        this.QuibService.deleteUser(order).subscribe(res => {
+          if (res) {
+            this.toastr.showSuccess("user deleted successfully", "user delete")
+            this.getOrderList()
+          }
+        })
+      },
+    });
+
   }
   getOrderList() {
     this.QuibService.getUserList().subscribe((data) => {
-      this.Quib_User =  data
+      this.Quib_User = data
       this.ngxLoader.stop();
     });
   }
-  markAsActive(id:number,Status:boolean){
+  markAsActive(id: number, Status: boolean) {
     this.ngxLoader.start();
-    this.QuibService.markAsActive(id,Status).subscribe(res => {
+    this.QuibService.markAsActive(id, Status).subscribe(res => {
       if (res) {
         this.toastr.showSuccess(" Status change successfully", "Status change")
         this.getOrderList()
